requirements.txt
pytest 
requests
yt-dlp
openai-whisper
torch
python-dotenv
google-genai
Django>=4.0
gunicorn
